Fix handling of underscores in class name and serializer-declaration
When a serializer was declared in a class that contains underscores, the
error message suggested a name that was different from the name that
was actually being generated for the serializer.

This resulted in either not being able to generate the source, or
when following the error message, having a compile error since the
referenced field didn't exist.

This commit changes the behaviour to follow the error message, i.e.
for a class 'Name_With_Underscore' the field-name is generated as
'_$name_With_UnderscoreSerializer' instead of
'_$nameWithUnderscoreSerializer'.