Makes `rebuild` take a void function.
It's not very clear whether a value returned by the `updates` parameter to `rebuild` means anything.
All examples use `=>` notation which returns the builder. The return type is currently `dynamic`, which *suggests* that the value is used for something, potentially something which cannot even be typed.

Marking the parameter as a `void` returning function makes it clear that you should not worry about the return value.